feat: verify ownership before update delete

+++ api/src/services/audioFiles/audioFiles.ts
@@ -4,6 +4,8 @@ import type {
   AudioFileRelationResolvers,
 } from 'types/graphql'
 
+import { ForbiddenError } from '@redwoodjs/graphql-server'
+
 import { db } from 'src/lib/db'
 
 export const audioFiles: QueryResolvers['audioFiles'] = () => {
@@ -24,19 +26,31 @@ export const createAudioFile: MutationResolvers['createAudioFile'] = ({
   })
 }
 
-export const updateAudioFile: MutationResolvers['updateAudioFile'] = ({
+const verifyOwnership = async (id) => {
+  if (await audioFile({ id })) {
+    return true
+  } else {
+    throw new ForbiddenError('You do not own this audio file')
+  }
+}
+
+export const updateAudioFile: MutationResolvers['updateAudioFile'] = async ({
   id,
   input,
 }) => {
+  await verifyOwnership(id)
+
   return db.audioFile.update({
     data: input,
     where: { id },
   })
 }
 
-export const deleteAudioFile: MutationResolvers['deleteAudioFile'] = ({
+export const deleteAudioFile: MutationResolvers['deleteAudioFile'] = async ({
   id,
 }) => {
+  await verifyOwnership(id)
+
   return db.audioFile.delete({
     where: { id },
   })
